Starting Over
Ok, so I believe part of the learning process is mistakes or errors you cant find until you walk away or in this case start all over and try it from a different approach. so in the last step, I didn't know why my style.css wasn't working with my HTML. Finally, I just threw it all out and started over. But this time I started with my navbar instead of my header. And for some strange reason the box-shadow I added worked. I never did on the first website and that's when I knew something was wrong but continued anyway but then more other parts weren't working also. some but not all so that's where I was lost of how and why. So here are the screenshots of that process.